requirements.txt
fastapi
uvicorn[standard]
celery
redis
sqlalchemy
pydantic
python-multipart
scikit-learn
pypdfium2
python-docx
python-magic
openai
httpx
transformers
python-dotenv
eventlet
torch